DW: improved table empty state, refactored

factoring sharedstates into their own componeonts


squash


sq


sq


fix rebase issue

sq
