+++ frontend/src/pages/distributedWorkloads/components/LoadingState.tsx
@@ -1,0 +1,8 @@
+import { Bullseye, Spinner } from '@patternfly/react-core';
+import React from 'react';
+
+export const LoadingState: React.FC = () => (
+  <Bullseye style={{ minHeight: 150 }}>
+    <Spinner />
+  </Bullseye>
+);
