+++ frontend/src/pages/distributedWorkloads/global/workloadStatus/DWWorkloadsTable.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Card, CardTitle, Bullseye, Spinner, Timestamp } from '@patternfly/react-core';
+import { Timestamp } from '@patternfly/react-core';
 import { Td, Tr } from '@patternfly/react-table';
 import EmptyStateErrorMessage from '~/components/EmptyStateErrorMessage';
 import { DistributedWorkloadsContext } from '~/concepts/distributedWorkloads/DistributedWorkloadsContext';
@@ -7,29 +7,23 @@ import { SortableData, Table } from '~/components/table';
 import { WorkloadKind } from '~/k8sTypes';
 import { getStatusInfo } from '~/concepts/distributedWorkloads/utils';
 import { WorkloadStatusLabel } from '~/pages/distributedWorkloads/components/WorkloadStatusLabel';
+import { NoWorkloadState } from '~/pages/distributedWorkloads/components/NoWorkloadState';
+import { LoadingState } from '~/pages/distributedWorkloads/components/LoadingState';
 
 export const DWWorkloadsTable: React.FC = () => {
   const { workloads } = React.useContext(DistributedWorkloadsContext);
 
   if (workloads.error) {
     return (
-      <Card isFullHeight>
-        <EmptyStateErrorMessage
-          title="Error loading distributed workloads"
-          bodyText={workloads.error.message}
-        />
-      </Card>
+      <EmptyStateErrorMessage
+        title="Error loading distributed workloads"
+        bodyText={workloads.error.message}
+      />
     );
   }
 
   if (!workloads.loaded) {
-    return (
-      <Card isFullHeight>
-        <Bullseye style={{ minHeight: 150 }}>
-          <Spinner />
-        </Bullseye>
-      </Card>
-    );
+    return <LoadingState />;
   }
 
   const columns: SortableData<WorkloadKind>[] = [
@@ -62,36 +56,37 @@ export const DWWorkloadsTable: React.FC = () => {
     },
   ];
 
+  if (!workloads.data.length) {
+    return (
+      <NoWorkloadState subTitle="Select another project or create a distributed workload in the selected project." />
+    );
+  }
   return (
-    <Card>
-      <CardTitle>Distributed Workloads</CardTitle>
-      <Table
-        enablePagination
-        data={workloads.data}
-        columns={columns}
-        emptyTableView={<>No distributed workloads match your filters</>}
-        data-id="workload-table"
-        rowRenderer={(workload) => {
-          const statusInfo = getStatusInfo(workload);
-          return (
-            <Tr key={workload.metadata?.uid}>
-              <Td dataLabel="Name">{workload.metadata?.name || 'Unnamed'}</Td>
-              <Td dataLabel="Priority">{workload.spec.priority}</Td>
-              <Td dataLabel="Status">
-                <WorkloadStatusLabel workload={workload} />
-              </Td>
-              <Td dataLabel="Created">
-                {workload.metadata?.creationTimestamp ? (
-                  <Timestamp date={new Date(workload.metadata.creationTimestamp)} />
-                ) : (
-                  'Unknown'
-                )}
-              </Td>
-              <Td dataLabel="Latest Message">{statusInfo.message}</Td>
-            </Tr>
-          );
-        }}
-      />
-    </Card>
+    <Table
+      enablePagination
+      data={workloads.data}
+      columns={columns}
+      data-id="workload-table"
+      rowRenderer={(workload) => {
+        const statusInfo = getStatusInfo(workload);
+        return (
+          <Tr key={workload.metadata?.uid}>
+            <Td dataLabel="Name">{workload.metadata?.name || 'Unnamed'}</Td>
+            <Td dataLabel="Priority">{workload.spec.priority}</Td>
+            <Td dataLabel="Status">
+              <WorkloadStatusLabel workload={workload} />
+            </Td>
+            <Td dataLabel="Created">
+              {workload.metadata?.creationTimestamp ? (
+                <Timestamp date={new Date(workload.metadata.creationTimestamp)} />
+              ) : (
+                'Unknown'
+              )}
+            </Td>
+            <Td dataLabel="Latest Message">{statusInfo.message}</Td>
+          </Tr>
+        );
+      }}
+    />
   );
 };
